feat: :sparkles: Split the option inputs into a separate component
